+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types.js";
 export * from "./engine.js";
 export * from "./errors.js";
+export * from "./validation.js";